78 and 79. Context, and hooks for Context - changing contextual values
The class 78 is poorly taught, the instructor creates the context and only on the end he realizes it does not work because it does not have a setter created for that context constant. In lesson 79 he basically only fix his mistake during the whole class, using a setState on the context provider value.