85. Custom hook
This one looks very bizarre, a lot parts of the code doesn't seem to work, as the IDE keeps telling some things are wrong, like the .catch after the await API call. Don't know if this is the course teaching bad practices or it is because of React different versions and new changes in the syntax.